refactor(server): peel cache routes into routers/cache.py (router-split, round-5 #51 / R5-25 PR16)
Fifth router peeled (merge-as-green off main after chat/PR15) — the last clean
group before the shared-helper blockers. The /api/cache group — cache-size report
+ clear (app/thumbnails/chromadb/waveforms/all) — moves into routers/cache.py with
its size helper (_dir_size_mb). Imports auth + config + state (embed single-flight
guard) + webguard (same-site guard) — no server import.

ROOT stays in server.py (shared by ingest/offload/etc.); the router derives the
same code directory from its own location (_CODE_ROOT = routers/cache.py →
routers/ → code root), verified == server.ROOT so waveforms/__pycache__ resolve
identically. The chromadb-clear rebuild guard imports the same state single-flight
object server exposes (asserted), so the 409 concurrency guard is preserved.

New boundary test pins leaf/ownership/mount + code-root parity + embed-guard
singleton sharing. Cache behaviour stays covered by test_hardening_round4.py.
852 passed.
