+++ server.py
@@ -159,10 +159,12 @@ from routers.admin import router as admin_router  # noqa: E402
 from routers.settings import router as settings_router  # noqa: E402
 from routers.projects import router as projects_router  # noqa: E402
 from routers.chat import router as chat_router  # noqa: E402
+from routers.cache import router as cache_router  # noqa: E402
 app.include_router(admin_router)
 app.include_router(settings_router)
 app.include_router(projects_router)
 app.include_router(chat_router)
+app.include_router(cache_router)
 
 ROOT = Path(__file__).parent
 # Built Svelte SPA (frontend/dist). Gitignored — produced by `npm run build`.
@@ -2206,124 +2208,8 @@ def reingest_media(
 
 # ── Cache Management ──────────────────────────────────────────────────────────
 
-def _dir_size_mb(p: Path) -> int:
-    total = 0
-    for f in p.rglob("*"):
-        try:
-            if f.is_file():
-                total += f.stat().st_size
-        except OSError:
-            # Skip un-stattable entries so a cache-size estimate never 500s. On
-            # Windows the HF cache's snapshots/ are symlinks into blobs/ that
-            # raise WinError 448 when symlink support is off — one bad link must
-            # not crash /api/cache/info.
-            continue
-    return round(total / 1048576)
-
-
-@app.get("/api/cache/info")
-def cache_info(
-    _tok: dict = Depends(require_scopes("videos_read")),
-):
-    """Show cache sizes.
-
-    Phase 16.2: the absolute `path` of each cache dir is intentionally NOT
-    returned — it leaked the operator's home/install layout to any videos_read
-    token. The dict key labels each cache; sizes/counts are what the UI needs.
-    """
-    caches = {}
-    # HuggingFace model cache
-    hf_cache = Path.home() / ".cache" / "huggingface"
-    if hf_cache.exists():
-        caches["huggingface"] = {"size_mb": _dir_size_mb(hf_cache)}
-    # Ollama models
-    ollama_dir = Path.home() / ".ollama" / "models"
-    if ollama_dir.exists():
-        caches["ollama"] = {"size_mb": _dir_size_mb(ollama_dir)}
-    # ChromaDB
-    if config.CHROMA_PATH.exists():
-        caches["chromadb"] = {"size_mb": _dir_size_mb(config.CHROMA_PATH)}
-    # Thumbnails
-    if config.THUMBNAILS_DIR.exists():
-        count = sum(1 for _ in config.THUMBNAILS_DIR.glob("*"))
-        caches["thumbnails"] = {"count": count, "size_mb": _dir_size_mb(config.THUMBNAILS_DIR)}
-    # Browser-playback proxies (generated by ingest for HEVC/ProRes sources)
-    if config.PROXIES_DIR.exists():
-        count = sum(1 for _ in config.PROXIES_DIR.glob("*.mp4"))
-        caches["proxies"] = {"count": count, "size_mb": _dir_size_mb(config.PROXIES_DIR)}
-    # Waveform peak cache
-    waveforms_dir = ROOT / "waveforms"
-    if waveforms_dir.exists():
-        count = sum(1 for _ in waveforms_dir.glob("*.json"))
-        caches["waveforms"] = {"count": count, "size_mb": _dir_size_mb(waveforms_dir)}
-    # Python __pycache__
-    pycache = ROOT / "__pycache__"
-    if pycache.exists():
-        caches["pycache"] = {"size_mb": _dir_size_mb(pycache)}
-    # Total
-    total_mb = sum(c.get("size_mb", 0) for c in caches.values())
-    return {"caches": caches, "total_mb": total_mb}
-
-
-@app.post("/api/cache/clear")
-def clear_cache(
-    request: Request,
-    target: str = Query("app", description="app|thumbnails|chromadb|waveforms|all"),
-    _tok: dict = Depends(require_scopes("videos_write")),
-):
-    """Clear caches. target: app (pycache+waveforms — cheap regenerables only),
-    thumbnails, chromadb, waveforms, all.
-
-    fable-audit round-5 #16: 'app' NO LONGER clears thumbnails. media.thumbnail_path
-    and the vision frame images (*_frame*.jpg) live in the thumbnails dir and are
-    DB-referenced — a casual 'clear app cache' used to blank the whole grid + frame
-    strips and need an hours-long --refresh. Thumbnails now clear only on an explicit
-    target of 'thumbnails' or 'all'."""
-    # fable-audit 2026-07-12 (#4): this is the only heavy mutating route missing the
-    # M14 guard — target=chromadb/all rmtree's the whole semantic index, and as a
-    # no-preflight simple POST a cross-site page could fire it under loopback-trust.
-    _assert_same_site(request)
-    import shutil
-    # #15: never rmtree the index out from under an active rebuild.
-    if target in ("chromadb", "all") and _embed_guard.active:
-        raise HTTPException(409, "語意索引重建進行中，請待完成後再清除")
-    cleared = []
-    if target in ("thumbnails", "all"):
-        thumbs = config.THUMBNAILS_DIR
-        if thumbs.exists():
-            files = list(thumbs.iterdir())
-            for f in files:
-                f.unlink(missing_ok=True)
-            cleared.append(f"thumbnails ({len(files)} removed)")
-    if target in ("app", "waveforms", "all"):
-        waveforms_dir = ROOT / "waveforms"
-        if waveforms_dir.exists():
-            files = list(waveforms_dir.glob("*.json"))
-            for f in files:
-                f.unlink(missing_ok=True)
-            cleared.append(f"waveforms ({len(files)} removed)")
-    if target in ("app", "all"):
-        pycache = ROOT / "__pycache__"
-        if pycache.exists():
-            shutil.rmtree(pycache, ignore_errors=True)
-            cleared.append("__pycache__")
-    if target in ("chromadb", "all"):
-        if config.CHROMA_PATH.exists():
-            try:
-                # #15: drop ignore_errors — a failed clear must surface, not be
-                # silently reported as done.
-                shutil.rmtree(config.CHROMA_PATH)
-            except OSError as exc:
-                raise HTTPException(500, "清除 chromadb 失敗：{0}".format(exc))
-            cleared.append("chromadb")
-        # #15: invalidate chromadb's in-process System cache so a subsequent rebuild
-        # is actually served, instead of the cached (now-deleted) index all session.
-        try:
-            import vectordb
-            vectordb.clear_client_cache()
-        except Exception:
-            pass
-    return {"ok": True, "cleared": cleared}
+# _dir_size_mb + the /api/cache/info + /api/cache/clear handlers moved to
+# routers/cache.py (R5-25 / #51), mounted via app.include_router below.
 
 
 # ── Export ────────────────────────────────────────────────────────────────────
